Do not use ActiveSupport::Notifications
Each instance of the TMS::Connection was getting added as an observer,
causing multiple log messages when there was more than one instance. It
would also have resulted in a memory leak for any long-lived process
that creates multiple instances of TMS::Client.

I think this was being done to get at the start and end time of the
request so we can log the duration. I'm just doing it in the TMS::Logger
middleware instead.